+++ Input/Wedge_Input.java
@@ -23,10 +23,10 @@ public class Wedge_Input {
         
         Output.legend_object();
             int obj_type_id=scn.nextInt();
-                float mass, mu, l, b;
-                float[] s0, u;
-                s0= new float[2];
-                u= new float[2];
+                double mass, mu, l, b;
+                double[] s0, u;
+                s0= new double[2];
+                u= new double[2];
                 mu=0;
                     System.out.println("Enter the mass of the object"); //We can replace the onject dynammically
                     mass=scn.nextFloat();
@@ -53,16 +53,16 @@ int i=1; //for object
                 case 2:
                 {
                     System.out.println("Enter the radius of the object");
-                    float r=scn.nextFloat();
+                    double r=scn.nextFloat();
                     obj.add(new Sphere(mass, mu, s0, u, i, obj_type_id,r));
                     break;
                 }
                 case 3:
                 {
                     System.out.println("Enter the radius of the object");
-                    float r=scn.nextFloat();
+                    double r=scn.nextFloat();
                     System.out.println("Enter the height of the object");
-                    float h=scn.nextFloat();
+                    double h=scn.nextFloat();
                     obj.add(new Cylinder(mass, mu, s0, u, i, obj_type_id, r, h));
                     break;
                 }
